fix legacy naming, refactor jit structure
